# backendApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gerenciamento de Clientes e Produtos Favoritos",
        "description": "Documentação da API para gerenciar autenticação, clientes e seus produtos favoritos.",
        "contact": {
            "name": "Suporte",
            "email": "suporte@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/clients/{client}/favorites": {
            "get": {
                "tags": [
                    "Favoritos"
                ],
                "summary": "Lista os favoritos de um cliente",
                "operationId": "listFavorites",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de favoritos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Favoritos"
                ],
                "summary": "Adiciona um produto aos favoritos de um cliente",
                "operationId": "addFavorite",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "ID do produto a ser favoritado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "description": "ID do produto da Fake Store API",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto adicionado aos favoritos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Favorite"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Product already favorited by this customer"
                    },
                    "404": {
                        "description": "Customer or product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clients/{client}/favorites/{product_id}": {
            "delete": {
                "tags": [
                    "Favoritos"
                ],
                "summary": "Remove um produto dos favoritos de um cliente",
                "operationId": "removeFavorite",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product removed from favorites"
                    },
                    "404": {
                        "description": "Favorite not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Favorite": {
                "properties": {
                    "id": {
                        "description": "ID do favorito",
                        "type": "integer",
                        "example": 1
                    },
                    "client_id": {
                        "description": "ID do cliente associado",
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "description": "ID do produto associado",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data de atualização",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "id": {
                        "description": "ID do produto",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Título do produto",
                        "type": "string",
                        "example": "Produto Teste"
                    },
                    "image": {
                        "description": "URL da imagem",
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "price": {
                        "description": "Preço do produto",
                        "type": "number",
                        "example": 29.99
                    },
                    "rating": {
                        "description": "Classificação do produto",
                        "type": "number",
                        "example": 4.5
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data de atualização",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}